Add files via upload
So... when i tried out the app it was working perfect. Just that one feature made me uncomfortable. That is, if we start the program from an outside console. the ¨exit()¨ command inmediatly close the console. of course this is not a problem if you are using an editor like VS. but i felt like creating a version that could be use by other people, like if they just want to check, they can download it in a pc and with nothing but just the console they can try out smoothly everything this little game has to offer.

Of course It's not necesary as it is just a simple personal. But At least i have learn a little more about funcitons and loops while trying to make things work the way i want.